minor cleanup (removing dublicate field definititions), extracting security related definitions to separate file.
